arreglo ingresos.php

+++ deleteIngresos.php
@@ -1,5 +1,15 @@
 <?php
-  include 'comprobar_conexion.php';
+  if(isset($_POST["id"])) {
+    $id = $_POST["id"];
+    $host = "localhost";
+    $username = "contador";
+    $password = "123456";
+    $dbname = "contadorvirtual";
+    $conn = mysqli_connect($host, $username, $password, $dbname);
+
+    if (!$conn) {
+      die("Connection failed: " . mysqli_connect_error());
+    }
 
     $sql = "DELETE FROM sueldo WHERE id=?";
     $stmt = $conn->prepare($sql);
@@ -12,5 +22,5 @@
     $stmt->execute();
 
     mysqli_close($conn);
- 
+  }
 ?>
